1-vanilla-agent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.3
python-dotenv==1.0.1
openai>=1.6.1
typing-extensions>=4.5.0
langchain-community>=0.0.1
